Initial emitter work
