+++ android/app/src/main/java/com/demo/activity/ActivityStarterModule.java
@@ -4,7 +4,6 @@ import android.app.Activity;
 import android.content.Intent;
 import android.net.Uri;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 
 import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.ReactNativeHost;
@@ -16,11 +15,6 @@ import com.facebook.react.bridge.ReactContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.WritableNativeArray;
-import com.facebook.react.bridge.WritableNativeMap;
-import com.facebook.react.modules.core.DeviceEventManagerModule;
-
-import java.util.HashMap;
-import java.util.Map;
 
 import javax.annotation.Nonnull;
 
@@ -29,16 +23,8 @@ import javax.annotation.Nonnull;
  */
 class ActivityStarterModule extends ReactContextBaseJavaModule {
 
-    private static DeviceEventManagerModule.RCTDeviceEventEmitter eventEmitter = null;
-
     ActivityStarterModule(ReactApplicationContext reactContext) {
         super(reactContext);
-    }
-
-    @Override
-    public void initialize() {
-        super.initialize();
-        eventEmitter = getReactApplicationContext().getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class);
     }
 
     /**
@@ -48,14 +34,6 @@ class ActivityStarterModule extends ReactContextBaseJavaModule {
     @Override
     public String getName() {
         return "ActivityStarter";
-    }
-
-    @Nullable
-    @Override
-    public Map<String, Object> getConstants() {
-        final Map<String, Object> constants = new HashMap<>();
-        constants.put("MyEventName", "MyEventValue");
-        return constants;
     }
 
     @ReactMethod
@@ -109,15 +87,13 @@ class ActivityStarterModule extends ReactContextBaseJavaModule {
                 CatalystInstance catalystInstance = reactContext.getCatalystInstance();
                 WritableNativeArray params = new WritableNativeArray();
                 params.pushString("Hello, JavaScript!");
+
+                // AFAIK, this approach to communicate from Java to JavaScript is officially undocumented.
+                // Use at own risk; prefer events.
+                // Note: Here we call 'alert', which shows UI. If this is done from an activity that
+                // doesn't forward lifecycle events to React Native, it wouldn't work.
                 catalystInstance.callFunction("JavaScriptVisibleToJava", "alert", params);
             }
         }
-    }
-
-    /**
-     * To pass an object instead of a simple string, create a {@link WritableNativeMap} and populate it.
-     */
-    static void emitEvent(@NonNull String message) {
-        eventEmitter.emit("MyEventValue", message);
     }
 }
