+++ android/app/src/main/java/com/demo/activity/EventEmitterModule.java
@@ -1,0 +1,67 @@
+package com.demo.activity;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.net.Uri;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+
+import com.facebook.react.ReactInstanceManager;
+import com.facebook.react.ReactNativeHost;
+import com.facebook.react.bridge.Callback;
+import com.facebook.react.bridge.CatalystInstance;
+import com.facebook.react.bridge.Promise;
+import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.bridge.ReactContext;
+import com.facebook.react.bridge.ReactContextBaseJavaModule;
+import com.facebook.react.bridge.ReactMethod;
+import com.facebook.react.bridge.WritableNativeArray;
+import com.facebook.react.bridge.WritableNativeMap;
+import com.facebook.react.modules.core.DeviceEventManagerModule;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import javax.annotation.Nonnull;
+
+/**
+ * Supports sending events to JavaScript.
+ */
+class EventEmitterModule extends ReactContextBaseJavaModule {
+
+    private static DeviceEventManagerModule.RCTDeviceEventEmitter eventEmitter = null;
+
+    EventEmitterModule(ReactApplicationContext reactContext) {
+        super(reactContext);
+    }
+
+    @Override
+    public void initialize() {
+        super.initialize();
+        eventEmitter = getReactApplicationContext().getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class);
+    }
+
+    /**
+     * @return the name of this module. This will be the name used to {@code require()} this module
+     * from JavaScript.
+     */
+    @Override
+    public String getName() {
+        return "EventEmitter";
+    }
+
+    @Nullable
+    @Override
+    public Map<String, Object> getConstants() {
+        final Map<String, Object> constants = new HashMap<>();
+        constants.put("MyEventName", "MyEventValue");
+        return constants;
+    }
+
+    /**
+     * To pass an object instead of a simple string, create a {@link WritableNativeMap} and populate it.
+     */
+    static void emitEvent(@NonNull String message) {
+        eventEmitter.emit("MyEventValue", message);
+    }
+}
