+++ android/app/src/main/java/com/demo/activity/ExampleActivity.java
@@ -1,9 +1,12 @@
 package com.demo.activity;
 
+import android.app.Activity;
 import android.os.Bundle;
 import android.support.annotation.CallSuper;
 import android.support.annotation.Nullable;
+import android.support.v4.app.ActivityCompat;
 import android.view.View;
+import android.widget.Toast;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactInstanceManager;
@@ -15,12 +18,8 @@ import com.facebook.react.bridge.WritableNativeArray;
 /**
  * Activity to start from React Native JavaScript, triggered via
  * {@link ActivityStarterModule#navigateToExample()}.
- *
- * This activity inherits ReactActivity in order to forward life cycle events to React Native.
- * This is not a requirement -- you can start any activity from React Native. See comments in
- * the trigger alert click listener for an explanation.
  */
-public class ExampleActivity extends ReactActivity {
+public class ExampleActivity extends Activity {
 
     @Override
     @CallSuper
@@ -38,11 +37,11 @@ public class ExampleActivity extends ReactActivity {
         findViewById(R.id.trigger_alert_button).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                ActivityStarterModule.emitEvent("Hello from " + ExampleActivity.class.getSimpleName());
+                EventEmitterModule.emitEvent("Hello from " + ExampleActivity.class.getSimpleName());
             }
         });
 
-        findViewById(R.id.call_javascript_button).setOnClickListener(new View.OnClickListener() {
+        findViewById(R.id.set_extra_message).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
                 MainApplication application = (MainApplication) ExampleActivity.this.getApplication();
@@ -53,13 +52,14 @@ public class ExampleActivity extends ReactActivity {
                 if (reactContext != null) {
                     CatalystInstance catalystInstance = reactContext.getCatalystInstance();
                     WritableNativeArray params = new WritableNativeArray();
-                    params.pushString("New extra message!");
+                    params.pushString("Set Extra Message was called!");
 
-                    // AFAIK, this approach is officially undocumented. Use at own risk; prefer events.
-                    // The JS function invoked calls 'alert'. If this activity didn't forward life cycle events
-                    // to React Native (which it does, due to inheriting ReactActivity), the alert dialog
-                    // would silently fail to show.
+                    // AFAIK, this approach to communicate from Java to JavaScript is officially undocumented.
+                    // Use at own risk; prefer events.
+                    // Note: Here we call 'setMessage', which does not show UI. That means it is okay
+                    // to call it from an activity that doesn't forward lifecycle events to React Native.
                     catalystInstance.callFunction("JavaScriptVisibleToJava", "setMessage", params);
+                    Toast.makeText(ExampleActivity.this, "Extra message was changed", Toast.LENGTH_SHORT).show();
                 }
             }
         });
