+++ index.ios.js
@@ -38,6 +38,7 @@ const exposedToJava = new ExposedToJava();
 BatchedBridge.registerCallableModule("JavaScriptVisibleToJava", exposedToJava);
 
 const activityStarter = NativeModules.ActivityStarter;
+const eventEmitterModule = NativeModules.EventEmitter;
 
 export default class ActivityDemoComponent extends Component {
   constructor(props) {
@@ -143,5 +144,5 @@ const styles = StyleSheet.create({
 
 AppRegistry.registerComponent('ActivityDemoComponent', () => ActivityDemoComponent);
 
-// const eventEmitter = new NativeEventEmitter(activityStarter);
-// eventEmitter.addListener(activityStarter.MyEventName, (params) => alert(params));
+const eventEmitter = new NativeEventEmitter(eventEmitterModule);
+eventEmitter.addListener(eventEmitterModule.MyEventName, (params) => alert(params));
